refactor: mover modelos Sequelize a infrastructure/adapters/outbound/persistence/entity
- mover 22 archivos de src/models/ a src/infrastructure/adapters/outbound/persistence/entity/
- actualizar importaciones en 23 archivos (repositorios, adapters, tests)
- corregir rutas relativas de DatabaseConfig en entidades
- alinear estructura con arquitectura hexagonal
- las entidades Sequelize ahora están correctamente ubicadas en capa de infraestructura
