+++ src/infrastructure/adapters/outbound/persistence/entity/ChannelConnection.model.js
@@ -1,5 +1,5 @@
 const { Model, DataTypes } = require('sequelize');
-const sequelize = require('../infrastructure/config/DatabaseConfig');
+const sequelize = require('../../../../config/DatabaseConfig');
 
 /**
  * Modelo unificado para todas las conexiones de canales de comunicación
