+++ src/infrastructure/adapters/outbound/persistence/entity/Employee.model.js
@@ -1,5 +1,5 @@
 const { Model, DataTypes } = require('sequelize');
-const sequelize = require('../infrastructure/config/DatabaseConfig');
+const sequelize = require('../../../../config/DatabaseConfig');
 
 class Employee extends Model {
     static associate(models){
