+++ src/infrastructure/adapters/outbound/persistence/entity/bankTrade.model.js
@@ -1,5 +1,5 @@
 const { Model, DataTypes } = require('sequelize');
-const sequelize = require("../infrastructure/config/DatabaseConfig");
+const sequelize = require("../../../../config/DatabaseConfig");
 const Country = require('./country.model');
 const Bank = require('./bank.model');
