+++ src/infrastructure/adapters/outbound/persistence/entity/type_account_bank.model.js
@@ -1,5 +1,5 @@
 const { Model, DataTypes } = require('sequelize');
-const sequelize = require('../infrastructure/config/DatabaseConfig');
+const sequelize = require('../../../../config/DatabaseConfig');
 
 class TypeAccountBank extends Model {}
