+++ src/infrastructure/adapters/outbound/persistence/entity/user.model.js
@@ -1,5 +1,5 @@
 const { Model, DataTypes } = require('sequelize');
-const sequelize = require('../infrastructure/config/DatabaseConfig');
+const sequelize = require('../../../../config/DatabaseConfig');
 const Customer = require('./customer.model');
 const UserType = require('./UserType.model');
